Refactor to remove noop_handler
The noop_handler is superfluous and can be handled by just setting the
current_log_handler to NULL.

It's unclear why this wasn't the original design. It probably wasn't for
MT safety, as the code is not MT safe to begin with[^1].

[^1]: In a situation where the current_log_handler is assigned to NULL
(by another thread) after it is checked but before it is called is just
as broken as if the function pointer was reassigned mid load. It's just
not safe to read/write it in such a context, so having it always point
to a "valid" fallback isn't useful.