Switch to a log handler singleton
Previously this would get created and dropped with each `Seat` instance.

Since that may somehow be possible in one process with the logind
backend, it would break logging when one of the two `Seat` instances was
dropped before the other. The logging configuration is global, so the
log setup should be global and through a singleton.