Improvements and fixes (#7)
* Update bindings to avoid &mut FFI_SEAT_LISTENER

This gets warnings in recent compilers regarding rust 2024 changes.

It's also not necessary anymore as of seatd commit 166feaea33 (released
in 0.6.0) [^1].

Notably: Someone building libseat-sys on a system with libseat 0.5.0 or
earlier using the `use_bindgen` feature will end up with a broken
libseat-sys 0.2.0.

[^1]: https://git.sr.ht/~kennylevinsen/seatd/commit/166feaea3394e00af14418e074ae090e31922f33

* Make elided lifetime explicit

* Fix typo

* Delete log_handler header file

The header file isn't used to generate bindings, there's no reason to
keep it separate.

* Refactor to remove noop_handler

The noop_handler is superfluous and can be handled by just setting the
current_log_handler to NULL.

It's unclear why this wasn't the original design. It probably wasn't for
MT safety, as the code is not MT safe to begin with[^1].

[^1]: In a situation where the current_log_handler is assigned to NULL
(by another thread) after it is checked but before it is called is just
as broken as if the function pointer was reassigned mid load. It's just
not safe to read/write it in such a context, so having it always point
to a "valid" fallback isn't useful.

* Remove userdata from log_handler

It is not used by the current code. It's just a NULL being passed
around.

* Set an explicit C standard version

The alternative is that in some years, some new C standard deprecates
something, GCC/Clang switch over to the new default, and start giving
warnings/errors.

* Don't expose log_handler.c APIs

* Add stderrlog to example

* Switch to a log handler singleton

Previously this would get created and dropped with each `Seat` instance.

Since that may somehow be possible in one process with the logind
backend, it would break logging when one of the two `Seat` instances was
dropped before the other. The logging configuration is global, so the
log setup should be global and through a singleton.

* Remove drop_preformatted_log_handler

Since the singleton will live for 'static once it is instantiated. drop
won't be called at any meaningful point, and is superfluous.

* Remove `LogHandler` entirely

Since the singleton doesn't expose any functions, doesn't implement drop
and doesn't hold any data, it doesn't need to be a singleton.

Use `Once` to ensure log setup happens only once.

* A number of opinionated changes to log_handler.c

* Make the `LogHandler` typedef a function type not a pointer type.
  This makes the typedef easier to read IMO. Additionally, explicily
  requiring * when it's used makes it obvious it's a pointer and
  statements like `type *v = NULL` look less confusing than
  `type v = NULL`;
* Inline `str_length`, only used once, should have been `static` anyway,
  and the name is misleading. It was an unnecessary indirection for a
  common idiom of calling a `*snprintf` function twice to get the
  expected size.
* Flatten `formatter_handler` removing the indentation around the core
  logic and returning early.
* Check for and return early if `malloc` fails. Possibly it would be
  better to propagate this fact back to rust but this is an improvement
  over invoking UB.

The style has been kept consistent (although I don't like it).